add plex entry in docker-compose.yml
plex is an alternative to jellyfin. It's closed-source and requires a subscription for the ability to stream from outside of your local network. I think it also supports automatic pre-transcoding of the video library to more compatible formats, but I've only tried it a little. Uncomment this block and add a PLEX_CLAIM token from https://account.plex.tv/claim